LPS-107004 Update tests
